+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
-## 0.10.0
+## 0.10.1
+### Enhancements
+* convenience methods `addSystems`/`addManagers` to add multiple `EntitySystem`s/`Manager`s
 
+## 0.10.0
 ### Breaking API Changes
 * entities are no longer simple `int`s and have been turned into an `extension type Entity(int)`. 
   Methods that previously expected `int entity` or `Iterable<int> entities>`